Refactor host view and asset details dashboards configuraton
