+++ packages/kbn-lens-embeddable-utils/attribute_builder/visualization_types/layers/xy_data_layer.ts
@@ -50,7 +50,7 @@ export interface XYLayerOptions {
   seriesType?: SeriesType;
 }
 
-interface XYLayerConfig {
+export interface XYLayerConfig {
   data: FormulaValueConfig[];
   options?: XYLayerOptions;
   /**
