+++ x-pack/plugins/infra/public/common/visualizations/index.ts
@@ -11,4 +11,5 @@ export type {
   HostsLensLineChartFormulas,
 } from './types';
 
+export * from './lens/dashboards';
 export { hostLensFormulas } from './constants';
