+++ x-pack/plugins/infra/public/common/visualizations/lens/dashboards/index.ts
@@ -1,0 +1,11 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { assetDetailsDashboards, type KPIChartProps } from './asset_details';
+export { hostsViewDashboards } from './hosts_view';
+export { AVERAGE_SUBTITLE, METRICS_TOOLTIP } from './translations';
+export { XY_MISSING_VALUE_DOTTED_LINE_CONFIG, KPI_CHART_HEIGHT } from './constants';
