+++ x-pack/plugins/infra/public/common/visualizations/types.ts
@@ -5,8 +5,19 @@
  * 2.0.
  */
 
+import type {
+  MetricLayerConfig,
+  XYLayerConfig,
+  XYReferenceLinesLayerConfig,
+} from '@kbn/lens-embeddable-utils';
 import { hostLensFormulas } from './constants';
 
 export type HostsLensFormulas = keyof typeof hostLensFormulas;
 export type HostsLensMetricChartFormulas = Exclude<HostsLensFormulas, 'diskIORead' | 'diskIOWrite'>;
 export type HostsLensLineChartFormulas = Exclude<HostsLensFormulas, 'hostCount'>;
+
+export type XYChartLayerParams =
+  | (XYLayerConfig & { type: 'visualization' })
+  | (XYReferenceLinesLayerConfig & { type: 'referenceLines' });
+
+export type MetricChartLayerParams = MetricLayerConfig & { type: 'visualization' };
