+++ x-pack/plugins/infra/public/components/asset_details/tabs/overview/kpis/kpi_grid.tsx
@@ -12,10 +12,10 @@ import type { TimeRange } from '@kbn/es-query';
 import { LensChart, TooltipContent } from '../../../../lens';
 import { buildCombinedHostsFilter } from '../../../../../utils/filters/build';
 import {
-  KPI_CHARTS,
+  assetDetailsDashboards,
   KPI_CHART_HEIGHT,
   AVERAGE_SUBTITLE,
-} from '../../../../../common/visualizations/lens/dashboards/host/kpi_grid_config';
+} from '../../../../../common/visualizations';
 
 interface Props {
   dataView?: DataView;
@@ -36,7 +36,7 @@ export const KPIGrid = React.memo(({ nodeName, dataView, timeRange }: Props) => 
 
   return (
     <EuiFlexGroup direction="row" gutterSize="s" data-test-subj="infraAssetDetailsKPIGrid">
-      {KPI_CHARTS.map(({ id, layers, title, toolTip }, index) => (
+      {assetDetailsDashboards.host.hostKPICharts.map(({ id, layers, title, toolTip }, index) => (
         <EuiFlexItem key={index}>
           <LensChart
             id={`infraAssetDetailsKPI${id}`}
