+++ x-pack/plugins/infra/public/pages/metrics/hosts/hooks/use_hosts_table.tsx
@@ -31,7 +31,7 @@ import { useUnifiedSearchContext } from './use_unified_search';
 import { useMetricsDataViewContext } from './use_data_view';
 import { ColumnHeader } from '../components/table/column_header';
 import { TABLE_COLUMN_LABEL } from '../translations';
-import { TOOLTIP } from '../../../../common/visualizations/lens/dashboards/host/translations';
+import { METRICS_TOOLTIP } from '../../../../common/visualizations';
 import { buildCombinedHostsFilter } from '../../../../utils/filters/build';
 
 /**
@@ -254,7 +254,7 @@ export const useHostsTable = () => {
         name: (
           <ColumnHeader
             label={TABLE_COLUMN_LABEL.cpuUsage}
-            toolTip={TOOLTIP.cpuUsage}
+            toolTip={METRICS_TOOLTIP.cpuUsage}
             formula={hostLensFormulas.cpuUsage.value}
           />
         ),
@@ -268,7 +268,7 @@ export const useHostsTable = () => {
         name: (
           <ColumnHeader
             label={TABLE_COLUMN_LABEL.normalizedLoad1m}
-            toolTip={TOOLTIP.normalizedLoad1m}
+            toolTip={METRICS_TOOLTIP.normalizedLoad1m}
             formula={hostLensFormulas.normalizedLoad1m.value}
           />
         ),
@@ -282,7 +282,7 @@ export const useHostsTable = () => {
         name: (
           <ColumnHeader
             label={TABLE_COLUMN_LABEL.memoryUsage}
-            toolTip={TOOLTIP.memoryUsage}
+            toolTip={METRICS_TOOLTIP.memoryUsage}
             formula={hostLensFormulas.memoryUsage.value}
           />
         ),
@@ -296,7 +296,7 @@ export const useHostsTable = () => {
         name: (
           <ColumnHeader
             label={TABLE_COLUMN_LABEL.memoryFree}
-            toolTip={TOOLTIP.memoryFree}
+            toolTip={METRICS_TOOLTIP.memoryFree}
             formula={hostLensFormulas.memoryFree.value}
           />
         ),
@@ -310,7 +310,7 @@ export const useHostsTable = () => {
         name: (
           <ColumnHeader
             label={TABLE_COLUMN_LABEL.diskSpaceUsage}
-            toolTip={TOOLTIP.diskSpaceUsage}
+            toolTip={METRICS_TOOLTIP.diskSpaceUsage}
             formula={hostLensFormulas.diskSpaceUsage.value}
           />
         ),
@@ -324,7 +324,7 @@ export const useHostsTable = () => {
         name: (
           <ColumnHeader
             label={TABLE_COLUMN_LABEL.rx}
-            toolTip={TOOLTIP.rx}
+            toolTip={METRICS_TOOLTIP.rx}
             formula={hostLensFormulas.rx.value}
           />
         ),
@@ -339,7 +339,7 @@ export const useHostsTable = () => {
         name: (
           <ColumnHeader
             label={TABLE_COLUMN_LABEL.tx}
-            toolTip={TOOLTIP.tx}
+            toolTip={METRICS_TOOLTIP.tx}
             formula={hostLensFormulas.tx.value}
           />
         ),
